Pipenv setup, project README, and other stuff

+++ main.py
@@ -2,15 +2,12 @@ import os
 import telebot
 from flask import Flask, request
 
-import config
+from config import token, heroku_webhook, welcome_message, HOST, PORT
 import spotify
 import ya_music
 
-bot = telebot.TeleBot(config.token)
+bot = telebot.TeleBot(token)
 bot.stop_polling()
-
-HOST = "0.0.0.0"
-PORT = os.environ.get('PORT', 8443)
 
 server = Flask(__name__)
 
@@ -19,18 +16,22 @@ sessionContext = {}
 
 @bot.message_handler(commands=["start"])
 def handle_start(message):
-    bot.send_message(message.from_user.id, config.welcome_message)
+    bot.send_message(message.from_user.id, welcome_message)
 
 
 @bot.message_handler(content_types=["text"])
 def handle_intent(message):
-    print(f'text handler {message}')
+    print(f"text handler {message}")
     process_command(message)
 
 
 def process_command(message):
     music_url = message.text
-    another_link = "cannot identify music provider or you send album(it's not supported now), only tracks"
+    another_link = """I don't understand this. I only support the following services (song links only for now):
+    - Spotify
+    - Я.Музыка (Yandex Music)
+    """
+
     try:
         if ya_music.is_ya_music(music_url):
             full_name = ya_music.get_full_track_name(music_url)
@@ -44,32 +45,34 @@ def process_command(message):
         bot.send_message(message.from_user.id, another_link)
 
 
-@server.route('/bot', methods=['POST'])
+@server.route("/bot", methods=["POST"])
 def post_message():
     req = request.stream.read().decode("utf-8")
-    print(f'post bot {req}')
+    print(f"post bot {req}")
     bot.process_new_updates([telebot.types.Update.de_json(req)])
-    return '/bot', 200
+    return "/bot", 200
 
 
-@server.route('/bot', methods=['GET'])
+@server.route("/bot", methods=["GET"])
 def get_message():
     req = request.stream.read().decode("utf-8")
-    print(f'get bot {req}')
+    print(f"get bot {req}")
     process_command(req)
-    return '/bot', 200
+    return "/bot", 200
 
 
-@server.route('/')
+@server.route("/")
 def webhook_handler():
     bot.remove_webhook()
-    bot.set_webhook(url=config.heroku_webhook)
+    bot.set_webhook(url=heroku_webhook)
     status_msg = f"i'm live. listening on {HOST}:{PORT}"
     return status_msg, 200
 
 
-bot.delete_webhook()
-bot.set_webhook(url=config.heroku_webhook)
-# uncomment for local testing and comment prev line
-# bot.polling(none_stop=True)
+if os.getenv("PYTHON_ENV") == "development":
+    bot.polling(none_stop=True)
+else:
+    bot.delete_webhook()
+    bot.set_webhook(url=heroku_webhook)
+
 server.run(host=HOST, port=PORT)
